Basic Drop Dash implementation
This commit adds in the Drop Dash as a new ability exclusive to Sonic. So far it replaces the insta-shield. Also takes into account the super form. No usage of specific spinning animations for the drop dash yet. But so far it should be working well and should feel very close, if not identical, to Sonic Mania.